iitj-chatbot-main/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langdetect==1.0.9
pydantic==2.8.2
httpx==0.27.2
slowapi==0.1.9
python-multipart==0.0.9
aiofiles==24.1.0
tiktoken==0.7.0
